Clean up stray functions and includes
