+++ include/overlay_01_021F3610.h
@@ -1,0 +1,12 @@
+#ifndef POKEHEARTGOLD_OV_01_021F3610_H
+#define POKEHEARTGOLD_OV_01_021F3610_H
+
+#include "global.h"
+
+void ov01_021F3B0C(VecFx32 *vec, int);
+BOOL ov01_021F3B30();
+void ov01_021F3B2C(int, int);
+int ov01_021F3B34(int);
+int ov01_021F3B44(int, u8 index);
+
+#endif // POKEHEARTGOLD_OV_01_021F3610_H
