+++ include/overlay_01_021F4704.h
@@ -4,5 +4,9 @@
 #include "field_system.h"
 
 void ov01_021F6304(FieldSystemUnkSub2C *a0);
+void ov01_021F630C(int, FieldSystemUnkSub2C *, s32 *);
+u8 ov01_021F6320(FieldSystemUnkSub2C *);
+u16 *ov01_021F65E4(FieldSystemUnkSub2C *, u8);
+u8 *ov01_021F65F0(FieldSystemUnkSub2C *, u8);
 
 #endif // POKEHEARTGOLD_OVERLAY_01_021F4704_H
