+++ include/unk_02034354.h
@@ -3,21 +3,22 @@
 
 #include "player_data.h"
 
-PlayerProfile *sub_02034818(u32 a0);
-void sub_02034AC0(SaveData *saveData, int a1);
 void sub_02034354(SaveData *saveData, int a1);
-u8 sub_02034870(int);
-int sub_02034750(int);
-void sub_02034434(void);
-int sub_02034780(void);
-void sub_020346E8(int);
-int sub_020347A0(void);
-BOOL sub_02034420(void);
-int sub_020346D4(void);
-int sub_02034730(int);
-int sub_02034714(int);
-void sub_0203476C(int);
-void sub_02034638(void);
 void sub_020343E4(void);
+BOOL sub_02034420(void);
+void sub_02034434(void);
+void sub_02034638(void);
+int sub_020346D4(void);
+void sub_020346E8(int);
+int sub_02034714(int);
+int sub_02034730(int);
+int sub_02034750(int);
+void sub_0203476C(int);
+int sub_02034780(void);
+int sub_020347A0(void);
+PlayerProfile *sub_02034818(u32 a0);
+u8 sub_02034870(int);
+s32 sub_020348F0();
+void sub_02034AC0(SaveData *saveData, int a1);
 
 #endif // POKEHEARTGOLD_UNK_02034354_H
