+++ include/unk_02034B0C.h
@@ -2,7 +2,13 @@
 #define POKEHEARTGOLD_UNK_02034B0C_H
 
 void sub_02034D8C(void);
-BOOL sub_02035650(void);
 BOOL sub_02034DB8(void);
+int sub_0203511C();
+u32 sub_02035150(u32 index);
+u8 sub_02035184();
+void sub_02035198();
+void sub_020351DC(u16 cursorPos, PlayerProfile *profile);
+BOOL sub_02035650(void);
+void sub_020356C0(u32 arg0);
 
 #endif // POKEHEARTGOLD_UNK_02034B0C_H
