+++ include/unk_02035900.h
@@ -1,18 +1,25 @@
 #ifndef POKEHEARTGOLD_UNK_02035900_H
 #define POKEHEARTGOLD_UNK_02035900_H
 
+#include "global.h"
+#include "safari_zone.h"
+
 BOOL sub_02036144(void);
+BOOL sub_020373B4(u16);
+BOOL sub_02037030(int, void *, int);
+int sub_02037454(void);
 BOOL sub_02037474(void);
 void sub_020374E4(void);
 void sub_02037504(void);
 BOOL sub_0203769C(void);
+void sub_020376D4(s32, SafariZoneAreaSet *areaSet);
+void sub_020376EC(int);
+BOOL sub_02037700();
+int sub_0203774C(u32 arg0);
+int sub_0203775C(u32 arg0);
+int sub_020378AC(int);
 void sub_020380F4(void);
 void sub_02038104(void);
-int sub_02037454(void);
-int sub_020378AC(int);
-BOOL sub_02037030(int, void *, int);
-void sub_020376EC(int);
-BOOL sub_020373B4(u16);
 int sub_02037824(int);
 BOOL sub_02037880(void);
