+++ include/unk_020379A0.h
@@ -4,8 +4,12 @@
 #include "heap.h"
 
 void sub_020379A0(enum HeapID heapID);
+BOOL sub_02037A10();
 void sub_02037AC0(u8 a0);
 BOOL sub_02037B38(u8 a0);
+void sub_02037B8C(u32, u8);
+s32 sub_02037BA0(s32, s32);
+void sub_02037BC8();
 void sub_02037BEC(void);
 u32 sub_02037C0C(u32 a0, u16 *a1);
 u16 *sub_02037C44(s32 a0);
