+++ include/unk_02058034.h
@@ -1,13 +1,19 @@
 #ifndef POKEHEARTGOLD_UNK_02058034_H
 #define POKEHEARTGOLD_UNK_02058034_H
 
+#include "field_system.h"
 #include "script.h"
 
+void sub_020580E4(FieldSystem *fieldSystem, s32 arg1, s32 arg2);
+void sub_02058124(FieldSystem *fieldSystem, u8, u32);
+void sub_02058164(u16 arg0);
+void sub_02058180();
 void sub_02058190(FieldSystem *fieldSystem);
 void sub_02058284(void);
 void sub_020582A8(void);
+void sub_02058720(void);
+int sub_020881C0(s16 *quantity, u16 a1);
 void sub_020588CC(void);
 void sub_020588DC(void);
-void sub_02058720(void);
 
 #endif // POKEHEARTGOLD_UNK_02058034_H
