+++ include/unk_02092B04.h
@@ -1,0 +1,9 @@
+#ifndef POKEHEARTGOLD_UNK_02092B04_H
+#define POKEHEARTGOLD_UNK_02092B04_H
+
+#include "global.h"
+#include "task.h"
+
+BOOL sub_02092B04(TaskManager *taskManager);
+
+#endif // POKEHEARTGOLD_UNK_02092B04_H
