+++ src/unk_02054E00.c
@@ -1,11 +1,6 @@
+#include "overlay_01_021F3610.h"
+#include "overlay_01_021F4704.h"
 #include "unk_02054E00.h"
-
-// functions still in asm:
-void ov01_021F630C(int, FieldSystemUnkSub2C *, s32 *);
-int ov01_021F3B44(int, u8);
-int ov01_021F3B30();
-int ov01_021F3B34(int);
-void ov01_021F3B2C(int, int);
 
 BOOL MapModel_IsHeadbuttTree(u32 mapModelNo) {
     if (mapModelNo == 0xd0) {
